[WIP] MAX_PATTERN_SIZE: Increase this to 64KiB
fio uses a mix of random bytes and a fixed pattern to generate
compressible data buffers. However this pattern can be pathological in
terms of the time taken to compress and/or decompress blocks of
data. So allow the user to input a largish (64KiB) file to use for the
compression data.

Right now this is a WIP and may interfere in bad way with other
options related to the patterns places in a buffer. Use with caution.