t/one-core-peak: Replace one-core-peak.sh with Python version
Having a Python script makes it easier to maintain going forward, so
replace the bash version with a Python version.

The Python version is intended to be functionally identical to the bash
version, except for a small number of error/info messages that have been
added/fixed/removed when appropriate.

Signed-off-by: Martin Oliveira <martin.oliveira@eideticom.com>